Rename NetCoreApp template to NetCoreRabbitMq
