+++ templates/Templates.sln
@@ -5,35 +5,26 @@ VisualStudioVersion = 15.0.26124.0
 MinimumVisualStudioVersion = 15.0.26124.0
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "content", "content", "{8F4CCB4E-CDEC-42F2-AB6D-831DBEE6CC70}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetCoreApp", "content\NetCoreApp\NetCoreApp.csproj", "{A4DFA0BE-17E3-4B10-9278-B14F210EBE20}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetCoreRabbitMq", "content\NetCoreRabbitMq\NetCoreRabbitMq.csproj", "{34CABEDD-4ADE-4BC7-901A-B0B9EFF31F23}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{34CABEDD-4ADE-4BC7-901A-B0B9EFF31F23}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{34CABEDD-4ADE-4BC7-901A-B0B9EFF31F23}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{34CABEDD-4ADE-4BC7-901A-B0B9EFF31F23}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{34CABEDD-4ADE-4BC7-901A-B0B9EFF31F23}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A4DFA0BE-17E3-4B10-9278-B14F210EBE20}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A4DFA0BE-17E3-4B10-9278-B14F210EBE20}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A4DFA0BE-17E3-4B10-9278-B14F210EBE20}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{A4DFA0BE-17E3-4B10-9278-B14F210EBE20}.Debug|x64.Build.0 = Debug|Any CPU
-		{A4DFA0BE-17E3-4B10-9278-B14F210EBE20}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{A4DFA0BE-17E3-4B10-9278-B14F210EBE20}.Debug|x86.Build.0 = Debug|Any CPU
-		{A4DFA0BE-17E3-4B10-9278-B14F210EBE20}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A4DFA0BE-17E3-4B10-9278-B14F210EBE20}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A4DFA0BE-17E3-4B10-9278-B14F210EBE20}.Release|x64.ActiveCfg = Release|Any CPU
-		{A4DFA0BE-17E3-4B10-9278-B14F210EBE20}.Release|x64.Build.0 = Release|Any CPU
-		{A4DFA0BE-17E3-4B10-9278-B14F210EBE20}.Release|x86.ActiveCfg = Release|Any CPU
-		{A4DFA0BE-17E3-4B10-9278-B14F210EBE20}.Release|x86.Build.0 = Release|Any CPU
-	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{A4DFA0BE-17E3-4B10-9278-B14F210EBE20} = {8F4CCB4E-CDEC-42F2-AB6D-831DBEE6CC70}
+		{34CABEDD-4ADE-4BC7-901A-B0B9EFF31F23} = {8F4CCB4E-CDEC-42F2-AB6D-831DBEE6CC70}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {91791A68-3320-45E8-B54C-C636DACBECF9}
 	EndGlobalSection
 EndGlobal
