+++ templates/content/NetCoreRabbitMq/Handlers/TestCommandHandler.cs
@@ -1,0 +1,13 @@
+﻿using System.Threading.Tasks;
+using Nybus;
+
+namespace NetCoreRabbitMq.Handlers
+{
+    public class TestCommandHandler : ICommandHandler<TestCommand>
+    {
+        public Task HandleAsync(IDispatcher dispatcher, ICommandContext<TestCommand> incomingCommand)
+        {
+            return Task.CompletedTask;
+        }
+    }
+}
