+++ templates/content/NetCoreRabbitMq/Handlers/TestEventHandler.cs
@@ -1,0 +1,13 @@
+﻿using System.Threading.Tasks;
+using Nybus;
+
+namespace NetCoreRabbitMq.Handlers
+{
+    public class TestEventHandler : IEventHandler<TestEvent>
+    {
+        public Task HandleAsync(IDispatcher dispatcher, IEventContext<TestEvent> incomingEvent)
+        {
+            return Task.CompletedTask;
+        }
+    }
+}
