+++ templates/content/NetCoreRabbitMq/NetCoreRabbitMq.csproj
