+++ templates/content/NetCoreRabbitMq/appsettings.Development.json
