+++ templates/content/NetCoreRabbitMq/appsettings.Production.json
