+++ templates/content/NetCoreRabbitMq/appsettings.json
