+++ templates/content/NetCoreRabbitMq/hostsettings.json
